requirements.txt
python-dotenv
beautifulsoup4
openai
chardet
markdownify
pydantic
